+++ CMakeLists.txt
@@ -59,11 +59,10 @@ endif()
 # files in your project. 
 #--------------------------------------------------------------
 
-add_library(HydroChrono STATIC "hydro_forces.cpp" "hydro_forces.h")
-add_executable(sphere_decay_demo "sphere_decay_demo.cpp")
-add_executable(sphere_decay_no_viz "sphere_decay_no_viz.cpp")
-add_executable(sphere_reg_waves_no_viz "sphere_reg_waves_no_viz.cpp")
-add_executable(rm3_demo "rm3_demo.cpp")
+add_library(HydroChrono STATIC "src/hydro_forces.cpp" "src/hydro_forces.h")
+add_executable(sphere_decay "demos/sphere/sphere_decay.cpp")
+add_executable(sphere_reg_waves "demos/sphere/sphere_reg_waves.cpp")
+add_executable(rm3_demo "demos/rm3/rm3_demo.cpp")
 #SWIG_ADD_LIBRARY(py_wec_chrono LANGUAGE python SOURCES hydro_forces_no_viz.i)
 
 target_compile_features(HydroChrono PUBLIC cxx_std_17)
@@ -82,17 +81,12 @@ set_target_properties(HydroChrono PROPERTIES
 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
-set_target_properties(sphere_decay_demo PROPERTIES 
+set_target_properties(sphere_decay PROPERTIES 
 	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
-set_target_properties(sphere_decay_no_viz PROPERTIES 
-	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
-	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
-	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
-
-set_target_properties(sphere_reg_waves_no_viz PROPERTIES 
+set_target_properties(sphere_reg_waves PROPERTIES 
 	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
@@ -107,9 +101,8 @@ set_target_properties(rm3_demo PROPERTIES
 #--------------------------------------------------------------
 
 target_link_libraries(HydroChrono ${LINK_LIBS})
-target_link_libraries(sphere_decay_no_viz HydroChrono)
-target_link_libraries(sphere_decay_demo HydroChrono)
-target_link_libraries(sphere_reg_waves_no_viz HydroChrono)
+target_link_libraries(sphere_decay HydroChrono)
+target_link_libraries(sphere_reg_waves HydroChrono)
 target_link_libraries(rm3_demo HydroChrono)
 #SWIG_LINK_LIBRARIES(py_wec_chrono wec_chrono ${PYTHON_LIBRARIES})
