+++ demos/rm3/geometry/float_global.obj
