+++ demos/rm3/geometry/plate.obj
