+++ demos/rm3/geometry/spar_global.obj
