+++ demos/rm3/geometry/torus.obj
