+++ demos/sphere/geometry/oes_task10_sphere.obj
