+++ demos/sphere/postprocessing/plot_sphere_decay.plt
